feat: add standalone install feedback component and support using in MDX
