+++ src/content/docs/apm/agents/php-agent/installation/php-agent-installation-overview.mdx
@@ -121,3 +121,5 @@ Installation procedures may be different for admins who install the PHP agent th
 * [Other partnership installation procedures](/docs/accounts-partnerships/partnerships/partner-based-installation/log-install-new-relic-partners)
 
 Not all partners support PHP agents.
+
+<InstallFeedback />
